requirements.txt
aioboto3>=9.4.0


asyncpg>=0.25.0
bcrypt>=3.2.0

email-validator>=1.1.3
fastapi>=0.75.0

fastapi_better_di>=0.3.1
passlib>=1.7.4

pydantic>=1.9.0

python-dotenv>=0.19.2

python-jose>=3.3.0
python-multipart>=0.0.5
SQLAlchemy>=1.4.31

sqlmodel>=0.0.6
starlette>=0.17.1
uvicorn>=0.17.5
